Code review suggestions and fixes
