+++ src/components/PublicMap/context/PublicFiltersContext.tsx
@@ -1,12 +1,12 @@
 import { createContext } from "react";
 import type { PublicMapDataRecordsQuery } from "@/__generated__/types";
-import type { PublicFiltersFormValue } from "@/types";
+import type { FilterField, PublicFiltersFormValue } from "@/types";
 
 export const PublicFiltersContext = createContext<{
   filtersDialogOpen: boolean;
-  setFiltersDialogOpen: React.Dispatch<
-    React.SetStateAction<boolean>
-  >;
+  setFiltersDialogOpen: React.Dispatch<React.SetStateAction<boolean>>;
+  filterFields: FilterField[];
+  setFilterFields: React.Dispatch<React.SetStateAction<FilterField[]>>;
   publicFilters: PublicFiltersFormValue[];
   setPublicFilters: React.Dispatch<
     React.SetStateAction<PublicFiltersFormValue[]>
@@ -20,6 +20,8 @@ export const PublicFiltersContext = createContext<{
 }>({
   filtersDialogOpen: false,
   setFiltersDialogOpen: () => null,
+  filterFields: [],
+  setFilterFields: () => [],
   publicFilters: [],
   setPublicFilters: () => [],
   records: [],
